requirements.txt
appdirs==1.4.4
argparse==1.4.0
asttokens==2.4.1
attrs==23.2.0
automat==22.10.0
axial-positional-embedding==0.2.1
brotli==1.0.9
buildtools==1.0.6
causal-conv1d==1.0.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
cmake==3.28.1
comm==0.2.1
constantly==23.10.4
cryptography==41.0.7
debugpy==1.8.0
decorator==5.1.1
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fsspec==2023.12.2
furl==2.1.3
gitdb==4.0.11
gitpython==3.1.40
gmpy2==2.1.2
greenlet==3.0.3
huggingface-hub==0.20.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==7.0.1
incremental==22.10.0
ipykernel==6.28.0
ipython==8.18.1
jedi==0.19.1
jinja2==3.1.2
joblib==1.3.2
jupyter-client==8.6.0
jupyter-core==5.7.0
lightning-utilities==0.10.0
lit==17.0.6
littleutils==0.2.2
local-attention==1.9.0
mamba-ssm==1.0.1
markupsafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.1
ninja==1.11.1.1
numpy==1.26.3
ogb==1.3.6
orderedmultidict==1.0.1
outdated==0.2.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
performer-pytorch==1.1.4
pexpect==4.9.0
pillow==10.0.1
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pygments==2.17.2
pyopenssl==23.2.0
pyparsing==3.1.1
pysocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyyaml==6.0.1
pyzmq==25.1.2
rdkit==2023.9.3
redo==2.0.4
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentry-sdk==1.39.1
setproctitle==1.3.3
setuptools==68.2.2
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sqlalchemy==2.0.25
stack-data==0.6.3
sympy==1.12
tensorboardx==2.6.2.2
threadpoolctl==3.2.0
tokenizers==0.15.0
torchmetrics==0.9.3
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
twisted==23.10.0
typing_extensions==4.7.1
tzdata==2023.4
urllib3==1.26.18
wandb==0.16.1
wcwidth==0.2.12
wheel==0.41.2
yacs==0.1.8
zipp==3.17.0
zope-interface==6.1
